Revert "TT-963 Move IDP list endpoint to config"
