+++ config/initializers/20_api_proxies.rb
@@ -1,8 +1,7 @@
 require 'originating_ip_store'
 
 Rails.application.config.after_initialize do
-  ida_frontend_client = Api::Client.new(CONFIG.ida_frontend_host, Api::ResponseHandler.new)
-  SESSION_PROXY = SessionProxy.new(ida_frontend_client, OriginatingIpStore)
-  config_api_client = Api::Client.new(CONFIG.config_api_host, Api::ResponseHandler.new)
-  CONFIG_PROXY = ConfigProxy.new(config_api_client)
+  IDA_FRONTEND_CLIENT = Api::Client.new(CONFIG.ida_frontend_host, Api::ResponseHandler.new)
+  SESSION_PROXY = SessionProxy.new(IDA_FRONTEND_CLIENT, OriginatingIpStore)
+  CONFIG_API_CLIENT = Api::Client.new(CONFIG.config_api_host, Api::ResponseHandler.new)
 end
